+++ src/Encamina.Enmarcha.AI.OpenAI.Azure/AzureOpenAIOptions.cs
@@ -4,6 +4,7 @@ using Azure.AI.OpenAI;
 
 using Encamina.Enmarcha.AI.OpenAI.Abstractions;
 
+using Encamina.Enmarcha.Core;
 using Encamina.Enmarcha.Core.DataAnnotations;
 
 namespace Encamina.Enmarcha.AI.OpenAI.Azure;
@@ -23,24 +24,29 @@ public sealed class AzureOpenAIOptions : OpenAIOptionsBase
     /// </summary>
     [Required]
     [Uri]
-    public Uri Endpoint { get; init; }
+    public required Uri Endpoint { get; init; }
 
     /// <summary>
     /// Gets the key credential used to authenticate to an LLM resource.
     /// This property is required if property <see cref="UseTokenCredentialAuthentication"/> is <c>false</c>.
     /// </summary>
-    [RequiredIf(nameof(UseTokenCredentialAuthentication), false)]
-    [NotEmptyOrWhitespace]
-    public string Key { get; init; }
+    [RequiredIf([nameof(UseTokenCredentialAuthentication), nameof(UseDefaultAzureCredentialAuthentication)], [false, false], failOnAnyCondition: false)]
+    public string? Key { get; init; }
 
     /// <summary>
     /// Gets a value indicating whether Token Credential authentication should be used.
-    /// If set to true, the value of the <see cref="Key"/> property is ignored.
+    /// If set to <see langword="true"/>, the value of the <see cref="Key"/> property is ignored.
     /// </summary>
     public bool UseTokenCredentialAuthentication { get; init; }
 
     /// <summary>
+    /// Gets a value indicating whether Default Azure Credential authentication should be used.
+    /// If set to <see langword="true"/>, the value of the <see cref="Key"/> property is ignored.
+    /// </summary>
+    public bool UseDefaultAzureCredentialAuthentication { get; init; } = false;
+
+    /// <summary>
     /// Gets the token credentials options to authenticate to an LLM resource.
     /// </summary>
-    public TokenCredentialsOptions TokenCredentialsOptions { get; init; }
+    public TokenCredentialsOptions? TokenCredentialsOptions { get; init; }
 }